requirements.txt
requests
feedparser
python-dateutil
openai>=1.0.0
pyyaml
